remove type cast

+++ 04-functions/main.go
@@ -10,5 +10,5 @@ func div(num float32, denom float32) float32 {
 	if denom == 0 {
 		return 0
 	}
-	return float32(num) / float32(denom)
+	return num / denom
 }
